refact: abstract folders; move gungi-data to data-services/abstract/data; rename gungi.dao.ts, gungi-data-model.ts, gungi-repository.ts
