--- server/src/data-services/abstract/data.ts
@@ -1,6 +1,0 @@
-export interface GungiData {
-  _id: string;
-  level: string;
-  loser: string;
-  winner: string;
-}
