+++ server/src/frameworks/data-services/mongo/mongo-data-services.service.ts
@@ -1,11 +1,11 @@
 import { Injectable } from '@nestjs/common';
-import { MongoGungiRepository } from './gungi-mongo-repository';
+import { GungiRepository } from './gungi-repository';
 import Gungi from 'src/domain/Gungi';
 import { IDataServices } from 'src/data-services/abstract/data-services.abstract';
 
 @Injectable()
 export class MongoDataServices implements IDataServices {
-  constructor(readonly gungi: MongoGungiRepository) {}
+  constructor(readonly gungi: GungiRepository) {}
 
   async findGungiById(id: string): Promise<Gungi | null> {
     return this.gungi.get(id);
